suggest: Move `SuggestionProvider` into its own module.
The `rs` module contains crate-internal types related to Remote
Settings. Since `SuggestionProvider` is part of the public interface,
and has non-trivial methods, let's give it its own module.
